refactor: rdflib.js to namespaces.js
