+++ app/utils/validation/add-field-valdiations-to-formNodesL.js
@@ -1,6 +1,6 @@
 import { Statement } from 'rdflib';
 import { getValidationSubjectsOnNode } from '../forking-store-helpers';
-import { EXT } from '../rdflib';
+import { EXT } from '../namespaces';
 import { FORM } from '@lblod/submission-form-helpers';
 import { createBlankNodeForValidation } from './create-blankNode-for-validation';
