+++ app/utils/validation/helpers.js
@@ -1,7 +1,7 @@
 import { getLocalFileContentAsText } from '../get-local-file-content';
 import { GRAPHS } from '../../controllers/formbuilder/edit';
 import { sym as RDFNode } from 'rdflib';
-import { EXT } from '../rdflib';
+import { EXT } from '../namespaces';
 import { FORM, RDF } from '@lblod/submission-form-helpers';
 
 export const validationGraphs = {
